gemini-mlops-production/requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
requests>=2.28.0
